refactor: Separate event on type
