+++ src/engine/events/event.hpp
@@ -51,49 +51,3 @@ public:
 private:
   Event &event;
 };
-
-class WindowResizeEvent : public Event
-{
-public:
-  WindowResizeEvent(int32_t width, int32_t height)
-  {
-    this->width = width;
-    this->height = height;
-  }
-
-  int32_t getWidth() { return width; }
-  int32_t getHeight() { return height; }
-
-  std::string toString() const override
-  {
-    std::stringstream ss;
-    ss << "WindowResizeEvent: " << width << ", " << height;
-    return ss.str();
-  }
-
-  EVENT_CLASS_TYPE(WindowResize);
-
-private:
-  int32_t width, height;
-};
-
-class MouseMoveEvent : public Event
-{
-public:
-  MouseMoveEvent(double x, double y) : X(x), Y(y) {}
-
-  double getX() { return X; }
-  double getY() { return Y; }
-
-  std::string toString() const override
-  {
-    std::stringstream ss;
-    ss << "MouseMoveEvent: " << X << ", " << Y;
-    return ss.str();
-  }
-
-  EVENT_CLASS_TYPE(MouseMove);
-
-private:
-  double X, Y;
-};
